+++ scripts/build_gitlab/win.ps1
@@ -1,4 +1,4 @@
-## Copyright 2009-2019 Intel Corporation
+## Copyright 2009-2020 Intel Corporation
 ## SPDX-License-Identifier: Apache-2.0
 
 md build
@@ -9,10 +9,7 @@ cmake --version
 cmake -L `
   -G $args[0] `
   -T $args[1] `
-  -D BUILD_OSPRAY_CI_TESTS=ON `
   -D BUILD_EMBREE_FROM_SOURCE=OFF `
-  -D BUILD_OIDN=ON `
-  -D INSTALL_IN_SEPARATE_DIRECTORIES=OFF `
   ../scripts/superbuild
 
 cmake --build . --config Release --target ALL_BUILD