+++ scripts/release/linux.sh
@@ -48,17 +48,8 @@ done
 ROOT_DIR=$PWD
 DEP_DIR=$ROOT_DIR/deps
 DEP_BUILD_DIR=$ROOT_DIR/build_deps
-OSPRAY_PKG_BASE=ospray-2.1.0.x86_64.linux
-OSPRAY_BUILD_DIR=$ROOT_DIR/build_release
-INSTALL_DIR=$ROOT_DIR/install/$OSPRAY_PKG_BASE
 THREADS=`nproc`
 
-#### Cleanup any existing directories ####
-
-rm -rf $DEP_DIR
-rm -rf $DEP_BUILD_DIR
-rm -rf $OSPRAY_BUILD_DIR
-rm -rf $INSTALL_DIR
 
 #### Build dependencies ####
 
@@ -72,6 +63,7 @@ cmake --version
 
 cmake \
   "$@" \
+  -D BUILD_JOBS=$THREADS \
   -D BUILD_DEPENDENCIES_ONLY=ON \
   -D CMAKE_INSTALL_PREFIX=$DEP_DIR \
   -D CMAKE_INSTALL_LIBDIR=lib \
@@ -87,8 +79,8 @@ cd $ROOT_DIR
 
 #### Build OSPRay ####
 
-mkdir -p $OSPRAY_BUILD_DIR
-cd $OSPRAY_BUILD_DIR
+mkdir -p build_release
+cd build_release
 
 # Clean out build directory to be sure we are doing a fresh build
 rm -rf *
@@ -103,45 +95,30 @@ export OpenImageDenoise_DIR=$DEP_DIR
 
 # set release settings
 cmake -L \
-  -D CMAKE_INSTALL_PREFIX=$INSTALL_DIR \
-  -D ISPC_EXECUTABLE=$DEP_DIR/bin/ispc \
   -D OSPRAY_BUILD_ISA=ALL \
-  -D OSPRAY_MODULE_DENOISER=ON \
-  -D OSPRAY_INSTALL_DEPENDENCIES=OFF \
+  -D ISPC_EXECUTABLE=$DEP_DIR/bin/ispc \
   -D OSPRAY_ZIP_MODE=ON \
-  -D CPACK_PACKAGING_INSTALL_PREFIX=/ \
+  -D OSPRAY_MODULE_DENOISER=ON \
+  -D OSPRAY_INSTALL_DEPENDENCIES=ON \
   -D CMAKE_INSTALL_INCLUDEDIR=include \
   -D CMAKE_INSTALL_LIBDIR=lib \
   -D CMAKE_INSTALL_DOCDIR=doc \
   -D CMAKE_INSTALL_BINDIR=bin \
   ..
 
-# build OSPRay
-make -j $THREADS install
+# create tar.gz
+make -j $THREADS preinstall
 
 # verify libs
-check_symbols $INSTALL_DIR/lib/libospray.so GLIBC   2 17 0
-check_symbols $INSTALL_DIR/lib/libospray.so GLIBCXX 3 4 19
-check_symbols $INSTALL_DIR/lib/libospray.so CXXABI  1 3 7
+check_symbols libospray.so GLIBC   2 17 0
+check_symbols libospray.so GLIBCXX 3 4 19
+check_symbols libospray.so CXXABI  1 3 7
 
-check_symbols $INSTALL_DIR/lib/libospray_module_ispc.so GLIBC   2 17 0
-check_symbols $INSTALL_DIR/lib/libospray_module_ispc.so GLIBCXX 3 4 19
-check_symbols $INSTALL_DIR/lib/libospray_module_ispc.so CXXABI  1 3 7
+check_symbols libospray_module_ispc.so GLIBC   2 17 0
+check_symbols libospray_module_ispc.so GLIBCXX 3 4 19
+check_symbols libospray_module_ispc.so CXXABI  1 3 7
 
-check_imf $INSTALL_DIR/lib/libospray.so
-check_imf $INSTALL_DIR/lib/libospray_module_ispc.so
+check_imf libospray.so
+check_imf libospray_module_ispc.so
 
-# copy dependent libs into the install
-INSTALL_LIB_DIR=$INSTALL_DIR/lib
-
-cp -P $DEP_DIR/lib/*ospcommon.so* $INSTALL_LIB_DIR
-cp -P $DEP_DIR/lib/*openvkl*.so* $INSTALL_LIB_DIR
-cp -P $DEP_BUILD_DIR/embree/src/lib/*embree*.so* $INSTALL_LIB_DIR
-cp -P $DEP_BUILD_DIR/oidn/src/lib/*OpenImage*.so* $INSTALL_LIB_DIR
-cp -P $DEP_BUILD_DIR/tbb/src/tbb/lib/intel64/gcc4.8/libtbb.so.* $INSTALL_LIB_DIR
-cp -P $DEP_BUILD_DIR/tbb/src/tbb/lib/intel64/gcc4.8/libtbbmalloc.so.* $INSTALL_LIB_DIR
-
-# tar up the results
-cd $INSTALL_DIR/..
-tar -caf $OSPRAY_PKG_BASE.tar.gz $OSPRAY_PKG_BASE
-mv *.tar.gz $ROOT_DIR
+make -j $THREADS package || exit 2